Put RNG stuff in a separate contract
